requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
astroid==2.12.13
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
celery==5.2.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dill==0.3.6
Django==4.1.5
django-celery-results==2.4.0
djangorestframework==3.14.0
drf-spectacular==0.25.1
frozenlist==1.3.3
idna==3.4
inflection==0.5.1
isort==5.11.4
jsonschema==4.17.3
kombu==5.2.4
lazy-object-proxy==1.9.0
mccabe==0.7.0
multidict==6.0.4
platformdirs==2.6.2
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pydantic==1.10.4
pylint==2.15.9
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyrsistent==0.19.3
python-dotenv==0.21.0
pytz==2022.7
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tomlkit==0.11.6
typing_extensions==4.4.0
uritemplate==4.1.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
yarl==1.8.2
